Update options.rpy
Proper building for the mod. To build it, select the mod from the renpy menu, go to "Build Distributions", and then do a Windows build.

You'll have to go fish the RPA out of the game folder in the thing it generates. Haven't worked out how to only generate the RPA yet.